# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Qnect API
  description: Qnect 백엔드 API (계정 · 프로필 · 관심사)
  version: v1
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
...
paths:
  /api/quizzes/me:
    get:
      tags:
      - 퀴즈
      summary: 내 퀴즈 조회
      description: 현재 활성화된 내 퀴즈를 정답 포함 편집 뷰로 반환한다. 아직 없으면 404(QUIZ_NOT_FOUND).
      operationId: getMine
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseQuizResponse"
    put:
      tags:
      - 퀴즈
      summary: 퀴즈 편집·저장
      description: 문항/보기를 직접 작성·수정·추가해 내 활성 퀴즈를 전체 교체한다(문항 3~5개).
      operationId: save
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSaveRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseQuizResponse"
  /api/profiles/me:
    get:
      tags:
      - 프로필
      summary: 내 프로필 조회
      description: 내 프로필을 반환한다. 미작성이면 기본정보만 채우고 profileCompleted=false 로 내려준다.
      operationId: getMine_1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseProfileResponse"
    put:
      tags:
      - 프로필
      summary: 프로필 작성·수정
      description: 프로필을 생성하거나 전체 교체한다. 온보딩·수정 공용.
      operationId: upsert
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileRequest"
        required: true
      responses:
        "400":
          description: 입력값 검증 실패
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseProfileResponse"
...
  /api/quizzes/owner/{ownerId}/attempts:
    post:
      tags:
      - 퀴즈
      summary: 퀴즈 응시·채점
      description: 친구 퀴즈에 답을 제출하고 채점 결과를 받는다. 프로필 완성 필요. 첫 풀기면 포인트가 적립된다.
      operationId: solve
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizAttemptRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseQuizResultResponse"
  /api/quizzes/me/generate:
    post:
      tags:
      - 퀴즈
      summary: AI 문제 초안 생성
      description: 내 프로필 기반으로 AI가 문제 초안 여러 개를 생성해 반환한다(저장하지 않음). 클라이언트가 '내 퀴즈 목록'에
        추가·편집한 뒤 PUT /me 로 저장한다. 기존 퀴즈를 덮어쓰지 않는다.
      operationId: generate
      parameters:
      - name: arg1
        in: query
        description: 퀴즈 종류(기본 FIRST_MEET)
        required: false
        schema:
          type: string
          default: FIRST_MEET
          enum:
          - FIRST_MEET
          - REMIND
      - name: arg2
        in: query
        description: "생성할 문제 수(1~5, 기본 3)"
        required: false
        schema:
          type: integer
          format: int32
          default: 3
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseAiQuizDraftResponse"
...
  /api/profiles/me/image:
    put:
      tags:
      - 프로필
      summary: 프로필 이미지(캐릭터) 설정
      description: 카탈로그의 캐릭터 식별자로 프로필 이미지를 설정한다. 무효 식별자는 400, 프로필 없으면 404.
      operationId: setImage
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileImageRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseImageResponse"
        "400":
          description: 무효 캐릭터 식별자(INVALID_INPUT)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseImageResponse"
        "404":
          description: 프로필 없음(PROFILE_NOT_FOUND)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseImageResponse"
...
  /api/friends:
    get:
      tags:
      - 친구
      summary: 내 친구 목록
      description: 수락된(ACCEPTED) 친구 전체를 반환한다.
      operationId: getFriends
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListFriendResponse"
    post:
      tags:
      - 친구
      summary: 친구 추가 요청
      description: 상대(addresseeId)에게 친구 요청을 보낸다. 중복 요청·자기 자신 요청은 거부.
      operationId: request
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseFriendResponse"
...
components:
  schemas:
    OptionInput:
      type: object
      properties:
        content:
          type: string
          description: 보기 텍스트
          maxLength: 255
          minLength: 0
        correct:
          type: boolean
          description: 정답 보기 여부
      required:
      - content
    QuestionInput:
      type: object
      properties:
        type:
          type: string
          description: 문항 종류(MULTIPLE/OX)
          enum:
          - MULTIPLE
          - OX
        content:
          type: string
          description: 질문
          maxLength: 255
          minLength: 0
        correctAnswer:
          type: string
          description: "정답(OX=O/X, 객관식=정답 보기 텍스트)"
          maxLength: 255
          minLength: 0
        required:
          type: boolean
          description: 필수 출제 여부
        options:
          type: array
          description: 객관식 보기(OX는 생략/빈 목록)
          items:
            $ref: "#/components/schemas/OptionInput"
      required:
      - content
      - correctAnswer
      - type
    QuizSaveRequest:
      type: object
      properties:
        questions:
          type: array
          description: 문항 목록(3~5개)
          items:
            $ref: "#/components/schemas/QuestionInput"
          maxItems: 5
          minItems: 3
      required:
      - questions
    ApiResponseQuizResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 요청 성공 여부
          example: true
        data:
          $ref: "#/components/schemas/QuizResponse"
          description: 성공 시 응답 데이터 (실패 시 null)
        error:
          $ref: "#/components/schemas/ErrorDetail"
          description: 실패 시 에러 정보 (성공 시 null)
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
          description: "에러 코드(예: INVALID_INPUT)"
          example: INVALID_INPUT
        message:
          type: string
          description: 에러 메시지
          example: 입력값이 올바르지 않습니다.
        fields:
          type: object
          additionalProperties:
            type: string
          description: 검증 실패 시 필드별 메시지 (해당 없으면 null)
    OptionView:
      type: object
      properties:
        optionId:
          type: integer
          format: int64
          description: 보기 ID
        content:
          type: string
          description: 보기 텍스트
        correct:
          type: boolean
          description: 정답 보기 여부
        seq:
          type: integer
          format: int32
          description: 보기 순서
    QuestionView:
      type: object
      properties:
        questionId:
          type: integer
          format: int64
          description: 문항 ID
        type:
          type: string
          description: 문항 종류
          enum:
          - MULTIPLE
          - OX
        content:
          type: string
          description: 질문
        correctAnswer:
          type: string
          description: "정답(OX=O/X, 객관식=정답 보기 텍스트)"
        required:
          type: boolean
          description: 필수 출제 여부
        seq:
          type: integer
          format: int32
          description: 문항 순서
        options:
          type: array
          description: 객관식 보기(OX는 빈 목록)
          items:
            $ref: "#/components/schemas/OptionView"
    QuizResponse:
      type: object
      properties:
        quizId:
          type: integer
          format: int64
          description: 퀴즈 ID
        type:
          type: string
          description: 퀴즈 종류
          enum:
          - FIRST_MEET
          - REMIND
        active:
          type: boolean
          description: 활성 여부
        questions:
          type: array
          description: 문항 목록(정답 포함)
          items:
            $ref: "#/components/schemas/QuestionView"
    ProfileRequest:
      type: object
      properties:
        school:
          type: string
          description: 학교명
          example: 국민대학교
          maxLength: 30
          minLength: 0
        gender:
          type: string
          description: 성별
          enum:
          - MALE
          - FEMALE
          example: MALE
        mbti:
          type: string
          description: MBTI(영문 4자)
          example: ENFP
          minLength: 1
          pattern: "^[A-Za-z]{4}$"
        drinkLevel:
          type: string
          description: 주량
          example: 소주 3잔
          maxLength: 20
          minLength: 0
        favoriteFood:
          type: string
          description: 좋아하는 음식
          example: 치킨
          maxLength: 30
          minLength: 0
      required:
      - drinkLevel
      - gender
      - mbti
      - school
    ApiResponseProfileResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 요청 성공 여부
          example: true
        data:
          $ref: "#/components/schemas/ProfileResponse"
          description: 성공 시 응답 데이터 (실패 시 null)
        error:
          $ref: "#/components/schemas/ErrorDetail"
          description: 실패 시 에러 정보 (성공 시 null)
    ProfileResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
          description: 사용자 ID
        name:
          type: string
          description: 이름
        age:
          type: integer
          format: int32
          description: 나이(생년월일로 계산)
        zodiac:
          type: string
          description: 띠(생년월일로 계산)
        school:
          type: string
          description: 학교명
        gender:
          type: string
          description: 성별
          enum:
          - MALE
          - FEMALE
        mbti:
          type: string
          description: MBTI
        drinkLevel:
          type: string
          description: 주량
        favoriteFood:
          type: string
          description: 좋아하는 음식
        imageUrl:
          type: string
          description: 프로필 이미지 URL
        publicCode:
          type: string
          description: QR/공유용 고유 코드
        profileCompleted:
          type: boolean
          description: 온보딩 완료 여부(프로필 작성됨)
...
    AnswerInput:
      type: object
      properties:
        questionId:
          type: integer
          format: int64
          description: 문항 ID
        answer:
          type: string
          description: "고른 답(객관식=보기 텍스트, OX=O/X)"
          maxLength: 255
          minLength: 0
      required:
      - answer
      - questionId
    QuizAttemptRequest:
      type: object
      properties:
        answers:
          type: array
          description: 문항별 답안(최대 문항 수만큼)
          items:
            $ref: "#/components/schemas/AnswerInput"
          maxItems: 5
          minItems: 1
      required:
      - answers
    AnswerResult:
      type: object
      properties:
        questionId:
          type: integer
          format: int64
          description: 문항 ID
        content:
          type: string
          description: 질문
        yourAnswer:
          type: string
          description: 내가 고른 답
        correctAnswer:
          type: string
          description: 정답
        correct:
          type: boolean
          description: 정답 여부
    ApiResponseQuizResultResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 요청 성공 여부
          example: true
        data:
          $ref: "#/components/schemas/QuizResultResponse"
          description: 성공 시 응답 데이터 (실패 시 null)
        error:
          $ref: "#/components/schemas/ErrorDetail"
          description: 실패 시 에러 정보 (성공 시 null)
    QuizResultResponse:
      type: object
      properties:
        attemptId:
          type: integer
          format: int64
          description: 응시 ID
        quizId:
          type: integer
          format: int64
          description: 퀴즈 ID
        score:
          type: integer
          format: int32
          description: 맞힌 개수
        total:
          type: integer
          format: int32
          description: 총 문항 수
        answers:
          type: array
          description: 문항별 채점 결과
          items:
            $ref: "#/components/schemas/AnswerResult"
    AiQuizDraftResponse:
      type: object
      properties:
        questions:
          type: array
          description: AI 생성 문제 초안 목록(미저장)
          items:
            $ref: "#/components/schemas/DraftQuestion"
    ApiResponseAiQuizDraftResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 요청 성공 여부
          example: true
        data:
          $ref: "#/components/schemas/AiQuizDraftResponse"
          description: 성공 시 응답 데이터 (실패 시 null)
        error:
          $ref: "#/components/schemas/ErrorDetail"
          description: 실패 시 에러 정보 (성공 시 null)
    DraftOption:
      type: object
      properties:
        content:
          type: string
          description: 보기 텍스트
        correct:
          type: boolean
          description: 정답 보기 여부
    DraftQuestion:
      type: object
      properties:
        type:
          type: string
          description: 문항 종류(MULTIPLE/OX)
          enum:
          - MULTIPLE
          - OX
        content:
          type: string
          description: 질문
        correctAnswer:
          type: string
          description: "정답(OX=O/X, 객관식=정답 보기 텍스트)"
        required:
          type: boolean
          description: 필수 출제 여부
        options:
          type: array
          description: "객관식 보기(2~4개, OX는 빈 목록)"
          items:
            $ref: "#/components/schemas/DraftOption"
    ApiResponseImageResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 요청 성공 여부
          example: true
        data:
          $ref: "#/components/schemas/ImageResponse"
          description: 성공 시 응답 데이터 (실패 시 null)
        error:
          $ref: "#/components/schemas/ErrorDetail"
          description: 실패 시 에러 정보 (성공 시 null)
    ImageResponse:
      type: object
      properties:
        imageUrl:
          type: string
          description: 설정된 프로필 이미지 URL(선택한 캐릭터의 이미지 URL)
    ProfileImageRequest:
      type: object
      properties:
        characterId:
          type: string
          description: 선택한 캐릭터 식별자
          example: character07
      required:
      - characterId
...
    FriendRequestDto:
      type: object
      properties:
        addresseeId:
          type: integer
          format: int64
      required:
      - addresseeId
    ApiResponseFriendResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 요청 성공 여부
          example: true
        data:
          $ref: "#/components/schemas/FriendResponse"
          description: 성공 시 응답 데이터 (실패 시 null)
        error:
          $ref: "#/components/schemas/ErrorDetail"
          description: 실패 시 에러 정보 (성공 시 null)
    FriendResponse:
      type: object
      properties:
        friendshipId:
          type: integer
          format: int64
        requesterId:
          type: integer
          format: int64
        addresseeId:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - PENDING
          - ACCEPTED
          - REJECTED
        acceptedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    ApiResponseListFriendResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 요청 성공 여부
          example: true
        data:
          type: array
          description: 성공 시 응답 데이터 (실패 시 null)
          items:
            $ref: "#/components/schemas/FriendResponse"
        error:
          $ref: "#/components/schemas/ErrorDetail"
          description: 실패 시 에러 정보 (성공 시 null)
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT